+++ src/PeriodicSystems.jl
@@ -33,7 +33,7 @@ import Base: +, -, *, /, \, (==), (!=), ^, isapprox, iszero, convert, promote_op
              hcat, vcat, hvcat, inv, show, lastindex, require_one_based_indexing, print, show, one, zero, eltype
 import MatrixPencils: isregular, rmeval
 import MatrixEquations: sylvd2!, luslv!
-import DescriptorSystems: isstable, horzcat, vertcat
+import DescriptorSystems: isstable, horzcat, vertcat, blockdiag, parallel, series, append
 import Polynomials: AbstractRationalFunction, AbstractPolynomial, poles, isconstant, variable, degree, pqs
 import Symbolics: derivative
 
@@ -48,7 +48,7 @@ export PeriodicTimeSeriesMatrix, PeriodicSwitchingMatrix, HarmonicArray, Fourier
 export isperiodic, isconstant, iscontinuous, islti, set_period
 export mb03vd!, mb03vy!, mb03bd!, mb03wd!, mb03kd! 
 export ps
-export psaverage, psc2d, psmrc2d, psteval
+export psaverage, psc2d, psmrc2d, psteval, psparallel, psseries, psappend, pshorzcat, psvertcat, psinv, psfeedback
 export ps2fls, hr2bt, hr2btupd, phasemat, ps2frls, DiagDerOp, ps2ls
 export pspole, pszero, isstable, psh2norm, pshanorm, pstimeresp, psstepresp
 export pdlyap, prdlyap, pfdlyap, pslyapd, pdlyaps!, pdlyaps1!, pdlyaps2!, pdlyaps3!, dpsylv2, dpsylv2!, pslyapdkr, dpsylv2krsol!, kronset!
@@ -81,7 +81,7 @@ include("psclyap.jl")
 include("pscric.jl")
 include("psdric.jl")
 #include("psstab.jl")
-#include("psops.jl")
+include("psops.jl")
 include("pmops.jl")
 include("psfutils.jl")
 include("psutils.jl")